Revert "Create database index for ES (#9800)"
This reverts commit b86549544be73c8614fda0a5dec05f2bcf474e7e.
