+++ packages/database/src/index.ts
@@ -1,4 +1,3 @@
-export type { LogEvent } from 'kysely'
 export { createDatabase, type Database } from './database'
 export { type Transaction } from './kysely'
 export { createTokenDatabase, type TokenDatabase } from './tokenDatabase'
