+++ packages/frontend/package.json
@@ -28,11 +28,11 @@
     "@elastic/elasticsearch": "^8.13.1",
     "@l2beat/backend-tools": "workspace:*",
     "@l2beat/config": "workspace:*",
-    "@l2beat/dal": "workspace:*",
     "@l2beat/database": "workspace:*",
     "@l2beat/shared": "workspace:*",
     "@l2beat/shared-pure": "workspace:*",
     "@l2beat/validate": "workspace:*",
+    "@l2beat/dal": "workspace:*",
     "@radix-ui/react-accordion": "^1.2.0",
     "@radix-ui/react-checkbox": "^1.1.0",
     "@radix-ui/react-collapsible": "^1.1.0",
@@ -64,6 +64,7 @@
     "fuzzysort": "^3.1.0",
     "gray-matter": "^4.0.3",
     "image-size": "^2.0.1",
+    "kysely": "^0.27.3",
     "lodash": "^4.17.21",
     "markdown-it": "^14.1.0",
     "next-themes": "^0.3.0",
