+++ packages/frontend/src/server/database.ts
@@ -1,9 +1,8 @@
-import '../dotenv'
-
-import type { Database, LogEvent } from '@l2beat/database'
+import type { Database } from '@l2beat/database'
 import { compiledToSqlQuery, createDatabase } from '@l2beat/database'
+import type { LogEvent } from 'kysely'
 import { env } from '~/env'
-import { createLogger } from './utils/logger'
+import { getLogger } from './utils/logger'
 
 let db: Database | undefined
 
@@ -15,7 +14,7 @@ export function getDb() {
           connectionString: env.DATABASE_URL,
           ssl: ssl(),
           ...pool(),
-          log: env.DATABASE_LOG_ENABLED ? makeDbLogger() : undefined,
+          log: env.DATABASE_LOG_ENABLED ? makeLogger() : undefined,
         })
       : createThrowingProxy()
   }
@@ -79,17 +78,14 @@ function pool() {
   }
 }
 
-function makeDbLogger() {
-  const indexPrefix =
-    env.DEPLOYMENT_ENV === 'staging' ? 'database-staging' : 'database-prod'
-  const logger = createLogger({ indexPrefix })
+function makeLogger() {
+  const logger = getLogger().for('Database')
 
   return (event: LogEvent) => {
     if (event.level === 'error') {
       logger.error('Query failed', {
         durationMs: event.queryDurationMillis,
         error: event.error,
-        source: 'frontend',
         sql: compiledToSqlQuery(event.query),
         ...(env.NODE_ENV === 'production'
           ? {
@@ -101,7 +97,6 @@ function makeDbLogger() {
     } else {
       logger.info('Query executed', {
         durationMs: event.queryDurationMillis,
-        source: 'frontend',
         sql: compiledToSqlQuery(event.query),
         ...(env.NODE_ENV === 'production'
           ? {
