+++ packages/frontend/src/server/utils/logger.ts
@@ -20,11 +20,6 @@ export function getLogger(): Logger {
     return logger
   }
 
-  logger = createLogger()
-  return logger
-}
-
-export function createLogger(opts?: { indexPrefix?: string }): Logger {
   const isLocal = env.NODE_ENV !== 'production'
 
   const loggerTransports: LoggerTransportOptions[] = [
@@ -47,7 +42,7 @@ export function createLogger(opts?: { indexPrefix?: string }): Logger {
     const options: ElasticSearchTransportOptions = {
       node: env.ES_NODE,
       apiKey: env.ES_API_KEY,
-      indexPrefix: opts?.indexPrefix ?? env.ES_INDEX_PREFIX,
+      indexPrefix: env.ES_INDEX_PREFIX,
       flushInterval: env.ES_FLUSH_INTERVAL,
     }
 
@@ -61,6 +56,7 @@ export function createLogger(opts?: { indexPrefix?: string }): Logger {
     logLevel: env.LOG_LEVEL,
     utc: isLocal ? false : true,
     transports: loggerTransports,
+    // TODO: make this configurable
     metricsEnabled: false,
   }
