+++ packages/public-api/src/utils/logger/createLogger.ts
@@ -12,10 +12,7 @@ import {
   type ElasticSearchTransportOptions,
 } from './elastic-search/ElasticSearchTransport'
 
-export function createLogger(
-  env: Env,
-  opts?: { indexPrefix?: string },
-): Logger {
+export function createLogger(env: Env): Logger {
   const isLocal = env.optionalString('DEPLOYMENT_ENV') === undefined
 
   const loggerTransports: LoggerTransportOptions[] = [
@@ -33,7 +30,7 @@ export function createLogger(
     const options: ElasticSearchTransportOptions = {
       node: env.string('ES_NODE'),
       apiKey: env.string('ES_API_KEY'),
-      indexPrefix: opts?.indexPrefix ?? env.string('ES_INDEX_PREFIX'),
+      indexPrefix: env.string('ES_INDEX_PREFIX'),
       flushInterval: env.optionalInteger('ES_FLUSH_INTERVAL'),
     }
